requirements.txt
########################################
# Application & runtime dependencies   #
########################################
python-dotenv>=1.0.1,<2
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
typing-extensions>=4.8,<5
prometheus-client>=0.20,<1
PyJWT>=2.8,<3
fastapi>=0.104,<1
uvicorn[standard]>=0.24,<1
pinecone-client>=3.0,<4
# Ingest providers (optional)
trafilatura>=1.6,<2
newspaper3k>=0.2,<1
yt-dlp>=2023.10,<2025
openai-whisper>=20230314,<20250101
telethon>=1.28,<2
langchain>=0.2.0,<0.3
langgraph>=0.2.30,<0.3
aiogram>=3.3,<4
aiofiles>=23.2,<24
boto3>=1.34,<2

########################################
# Dev / QA tooling                     #
########################################
pytest>=8.2,<9
pytest-asyncio>=0.23,<1
ruff>=0.6,<0.7
mypy>=1.10,<2
